restructuring files and adding setup.py for the packaging
