+++ bin/escli
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+import elsticsearchcli
+
+def main():
+    return elsticsearchcli.main()
+
+if __name__ == '__main__':
+    main()
